# Improve `CpuInfo` performances