--- a/Cargo.toml
+++ b/Cargo.toml
@@ -25,8 +25,13 @@ flate2 = { version = "1", optional = true }
 backtrace = { version = "0.3", optional = true }
 
 [dev-dependencies]
+criterion = "0.3"
 procinfo = "0.4.2"
 failure = "0.1"
 
 [package.metadata.docs.rs]
 all-features = true
+
+[[bench]]
+name = "cpuinfo"
+harness = false