--- a/src/meminfo.rs
+++ b/src/meminfo.rs
@@ -378,7 +378,9 @@ impl Meminfo {
             file_huge_pages: map.remove("FileHugePages"),
         };
 
-        assert!(!cfg!(test) || map.is_empty(), "meminfo map is not empty: {:#?}", map);
+        if cfg!(test) {
+            assert!(map.is_empty(), "meminfo map is not empty: {:#?}", map);
+        }
 
         Ok(meminfo)
     }