@eminence 2022-03-15
The CI failure is related to MSRV nonsense.  I'm in the process of bumping it via #170 so I think my recommendation is to wait for that to be merged and then rebase on top of the latest master