@macisamuele 2022-03-18
> The CI failure is related to MSRV nonsense. I'm in the process of bumping it via https://github.com/eminence/procfs/pull/170 so I think my recommendation is to wait for that to be merged and then rebase on top of the latest master

I've seen the failure and it is very odd, I've just rebased after #170 got merged, if the build fails again I'll be looking more into it

> I dunno what I was thinking when when I wrote the original version

lol, don't worry about that. The library is great and has very nice and easy to use APIs. 😊