@eminence 2022-03-18
Ah, I now understand the CI problem:  the failure is in the crate `half`, which is brought in my the new `criterion` dev-dependency.  So a normal "cargo build" is OK, but a "cargo test" will fail on 1.42.  You can fix this by pinning `half` to an older version:

    cargo +1.42 update -p half --precise 1.7.1 

(though note: we're very likely to bump MSRV again, see #171 )