@eminence 2022-04-05
Now that https://github.com/eminence/procfs/pull/171 is merged, can you rebase on the latest master branch please?